+++ packages/widgetbook_golden_test_alchemist/lib/widgetbook_golden_test_alchemist.dart
@@ -1,0 +1,4 @@
+export 'src/run_alchemist_widgetbook_golden_tests.dart';
+export 'package:widgetbook_golden_test_core/src/widgetbook_golden_tests_properties.dart';
+export 'package:widgetbook_golden_test_core/src/widgetbook_golden_test_builder.dart';
+export 'package:widgetbook_golden_test_core/src/golden_play_action.dart';