+++ packages/widgetbook_golden_test/CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Unreleased
-- Split into widgetbook_golden_test and widgetbook_golden_test_core.
+### Added
+- Add `constraints` property to `WidgetbookGoldenTestBuilder`.
+
+### Fixed
+- Include Scaffold in the widget tree for golden tests.
+
+### Updated
 - Bumped minimal widgetbook version required to 3.20.2.
+
+### Others
+- Split into widgetbook_golden_test and widgetbook_golden_test_core.
 
 ### Breaking changes
 - Deleted 'create_golden_test.dart' file. Its functionality has been moved to `WidgetbookGoldenFlutterTestRenderer`, `WidgetbookGoldenTestGenerator` and `goldenTestZoneRunner`.